requirements.txt
pyYAML
python-gitlab